scripts/python/requirements.txt
psycopg2-binary
faker
